Implement FloatEq on arrays.
This is a working implementation, but assert messages are going to be
broken for relative epsilon comparisons. There is an open design
question as to whether the rel_epsilon method ought to be changed to
reflect the fact that it is used to display output and may not be useful
by itself.