Introduce *_all checks.
Add `abs_all`, `rel_all` and `ulps_all` checks, which take a single
epsilon value and use it uniformly when checking every field. This is
the behaviour that FloatEq used to express, but that has now been
changed to use a structured epsilon value that works on a per field
basis. The FloatEqAll and FloatEqAllDebug traits have been added
to accomodate this.

This change is more painful than extending the checks to add the new
behaviour. However, I believe it is important to make this break now,
since the new FloatEq is the more general case, and there are some types
that will not want to implement FloatEqAll.