Added the FloatUlps trait
The rest of the traits have been reimplemented in terms of this trait,
which explicitly links a type to its ULPs representation. It allows for
a much natural `Ulps<T>` alias and paves the way for
`#[derive(FloatUlps)]` and hence derive macros for the rest of the
traits which rely upon it.