Derive support for all traits
Added the float_eq_derive crate, which provides custom derive macros
for every float_eq trait. The float_eq attribute is used to provide
additional configuration. FloatUlps, FloatEq, FloatDiff and FloatEqDebug
all require an `ulps = "MyTypeUlps"` option to be specified, with the
name of the relevant Ulps type, FloatEqAll and FloatEqAllDebug require
`all_epsilon = "EpsilonType"`. In keeping with the philosophy of the
rest of the crate's design, these must be explicitly given rather than
defaulted or deduced. This should also make it easier to search for.

- Derived Ulps types share the visibility of their parent type.
- Generic types are not yet derivable.