Reworked traits to better convey intended usage.
The intention is that the trait names and capabilities now match the
names of the features that they enable when you implement them,
hopefully reducing the friction of extending the library over new types.

The `FloatEqUlps` trait has now become `FloatEqUlpsEpsilon`, and a new
`FloatEqDebugUlpsDiff` trait added to differentiate between epsilon
parameters and ULPs based difference calculations. The `FloatEqDebug`
and `FloatEqAllDebug` traits have been renamed `AssertFloatEq` and
`AssertFloatEqAll` to match what needs to be implemented to allow types
to be used in those constructs. `FloatDiff` has been rolled into
`AssertFloatEq` and its methods prefixed with `debug_`, since it did not
provide the general difference calculations that its name implied.

The `#[derive]` macro tests are now better categorised and much more
comprehensive.