Added more relative checking modes.
The "rel" checking mode scales epsilon to the precision of the largest
operand. However, there are times when it would be more appropriate to
use the smaller of the two operands (as suggested by Knuth), or to
scale the threshold to the precision of the first or second operand,
regardless of size (for example if you're testing against a value
specifically calibrated for a unit test).

Hence, you may now use the following relative modes, given an equality
check between two operands a and b:

- rmax: epsilon is relative to max(a, b)'s precision
- rmin: epsilon is relative to min(a, b)'s precision
- r1st: epsilon is relative to a's precision
- r2nd: epsilon is relative to b's precision

This commit also contains a thorough pass over the unit tests, in
particular ensuring that comparing containers of different shapes does
not give false positives.