Rewrote and reorganised the documentation.
This new documentation is based on the https://diataxis.fr/ framework. An mdbook has added to contain articles that don't fit into the API documentation. A lot of the "how to" material has been factored out of the doc comments into book articles and an introductory floating point tutorial has been added.