API and docs now refer to tolerance/tol.
The API and the documentation have been changed to refer to tolerance/tol rather than epsilon/eps/margin/max_diff, since it is the mathematical term of art. This will aid user understanding and reduce ambiguity. This change wasn't made lightly, since I am generally averse to changing an API in such a big way, but we're still a 0.x version and I do not anticipate it happening again on the way to 1.0.0.

The version number will be bumped up to 0.6.0 for the next release accordingly.