Allow beta branch builds to fail for now.
The compiler on the beta branch emits errors for the derive implementations in a different order from the current stable compiler. They seem to be ordered more logically now, but in order for the build not to be marked as broken, I'll allow them to fail until the switchover, at which point we can fix the test cases and turn them back on again.